ConfigCatClientIntegrationTests
